requirements.txt
coinbase-advanced-py
matplotlib
mypy
numpy
pyqt5
python-dateutil
scipy
smtplib
sphinx
sphinx-rtd-theme
